Stable Build
Changed the inspectors and editors to be more readable, created backend framework for future updating as things are changed.

Made it possible to define static prefab sets in data sets, this to allow future manipulation of submeshes for objects that will be interactable/movable in a scene. Causes increase in batches as it counts each submesh object as its own but setpasses remain the same.

Added a bunch of backend API and other functionalities for programmers to use tool in scripts.