Added Observer patterns
To fail safe order of execution the merger tool component now creates an observer pattern if a data packet with its ID does not exist and will auto update itself if a packet is created for it.

This to ensure users will get packets even if they create things out of order.