Exponential vert growth fixed.
Noticed that the number of vertices in a batched mesh group would grow exponentially because it added the parent objects mesh onto the combined mesh each time for non-static groups.

It's a bit of a quick/dirty fix but made it so the mesh merger clears away the parent mesh to ensure it doesn't get added into the merger each time.

Mesh vertex counts gone down quite a bit now because of it.